# openapi.yaml
components:
  schemas:
    ErrorMessage:
      properties:
        detail:
          description: Error information
          title: Detail
          type: string
      required:
      - detail
      title: ErrorMessage
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: crypto_market_trader
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/v1/trades/{side}/{quantity}:
    get:
      operationId: excute_trade_api_v1_trades__side___quantity__get
      parameters:
      - in: path
        name: side
        required: true
        schema:
          title: Side
          type: string
      - in: path
        name: quantity
        required: true
        schema:
          title: Quantity
          type: number
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
          description: Not Found
          title: Not Found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
          description: Service Unavailable
          title: Service Temporarily Unavailable
      summary: Excute Trade
      tags:
      - Trading Orders Endpoints